ProductsAPI
The ProductsAPI project performs the operations of listing the products, fetching their details, creating, updating and deleting them. Product data is retrieved from the database with Entity Framework Core.
With ASP.Net Core Identity, users are registered and logged in, and JWT (JSON Web Token) is generated for secure access of users.